phase-3/backend/pyproject.toml
[project]
name = "todolist-pro-api"
version = "1.0.0"
description = "FastAPI Tasks CRUD API for TodoList Pro"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Tayyab Fayyaz" }
]
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.0.0",
    "uuid6>=2024.1.12",
    "pyjwt>=2.8.0",
    "uvicorn[standard]>=0.32.0",
    "alembic>=1.14.0",
    "greenlet>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "alembic/*"]
